medical_imaging_pipeline/requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Transformers for text processing
transformers>=4.30.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# API deployment
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0

# Medical imaging specific
albumentations>=1.3.0
opencv-python>=4.8.0
